+++ src/createApolloClient.js
@@ -1,12 +1,13 @@
 import { InMemoryCache, ApolloClient } from '@apollo/client';
 
-const createClient = ({ url, tenant, token }) => (new ApolloClient({
-  uri: `${url}/graphql`,
+const createClient = ({ tenant, token, url }) => (new ApolloClient({
+  cache: new InMemoryCache(),
+  credentials: 'include',
   headers: {
     'X-Okapi-Tenant': tenant,
-    'X-Okapi-Token': token,
+    ...(token && { 'X-Okapi-Token': token }),
   },
-  cache: new InMemoryCache(),
+  uri: `${url}/graphql`,
 }));
 
 export default createClient;